Resolves multiple issues when calling `azd` extension tool (#203)
Resolves multiple issues when azd is invoked from MCP server

Adds additional command parameters for cwd, environment and learn to provide additional context and description to the calling agent/LLM .

Enables the agent/LLM to learn more about the best practices for azd usage and guidelines in additional to the limited tool description.

Returns the azd best practices guide when the learn parameter has been set.

Encourages the agent/LLM to run long running operations like provision, deploy, up and down in a terminal instead of within the tool so users get incremental feedback. Today the MCP protocol does not support incremental streaming response feedback that can be handled. At some point in the future when better streaming support is available, we can re-visit this implementation.