Look for the McpServerToolAttribute in the right place (#388)
* Look for McpServerToolAttribute in the right place

Problem:

The `ToolOperations` type is responsible for finding all the relevant commands and exposing them as MCP tools. As part of this process it looks for the `McpServerToolAttribute` on implementations of `IBaseCommand.GetCommand` method and uses the metadata to fill out the `Tool.Annotations` data structure. The problem is that this isn't where the `McpServerToolAttribute` is used; all commands place it on their implementation of the `IBaseCommand.ExecuteAsync` method.

Since the code is looking in the wrong place the annotations are never discovered and can't be passed along to MCP clients.

Solution:

Update the code to look for the attribute on the `ExecuteAsync` method instead of `GetCommand`.

The `ToolOperations.GetAllTools` test has been updated to validate that we now set `Tool.Annotations` for every tool. It is safe to make this assertion since every tool is expected to have at least a title.

I also ran the MCP server and connected to it with a client to verify that the annotations were coming through as expected.

* Update CHANGELOG.md

Update CHANGELOG.md.