Proxy tool per service and initial support for remote servers (#366)
In this PR we are experimenting with an approach where instead of exposing a single azure tool we expose a tool per service area like storage and keyvault.

Internally the same proxy approach and tool narrowing approach is applied but gives users more details on what tools/services are exposed which also comes with more user controls on allowing tools to be invoked.

This POC also includes a pathway to enable support for remote MCP servers. The newly available Microsoft Learn is a remote MCP server and is mapped to and exposed as the "learn" tool from Azure MCP and child requests are forwarded to the remote MCP learn tool like any of our other included MCP tools.